requirements.txt
streamlit
requests
beautifulsoup4
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv
supabase
